Began overhaul of new logic gate drawing system
We're not yet back to original capabilities, so I'm branching for now even though I'm the only one working on the project (single tear). The new system will be canvas-based and completely agnostic of the fact that the project uses React, aside from a single file with React bindings (LogicGateSchematic.tsx), so it should be easy to rip for other projects. It uses a plain javascript object to represent a collection of gates with linkages between them and a modular strategy pattern to render it to a canvas. The new system should be faster and more flexible, although adding mouse interactions will require some shenanigans.