Minor refactor and corrected positioning of schematics
In order to make the canvas cooperate with CSS, I had to put it in a div, set it to fixed positioning, and then dynamically set the position, width, and height to match the div on render and on window resize. I get that having to redraw the canvas on resize might be undesirable in certain situations, but wanting to set a canvas size with css has got to be a common enough use case that they really should implement a standard way of doing it because this was just excessive.